+++ app/Console/Commands/Jav/XCityIdol.php
@@ -3,8 +3,8 @@
 namespace App\Console\Commands\Jav;
 
 use App\Jobs\XCityIdolFetchItem;
+use App\Services\Jav\XCityIdolService;
 use App\Services\TemporaryUrlService;
-use App\Services\XCityIdolService;
 use Illuminate\Console\Command;
 
 class XCityIdol extends Command
