+++ app/Console/Commands/Jav/XCityIdols.php
@@ -3,8 +3,8 @@
 namespace App\Console\Commands\Jav;
 
 use App\Jobs\XCityIdolFetchItems;
+use App\Services\Jav\XCityIdolService;
 use App\Services\TemporaryUrlService;
-use App\Services\XCityIdolService;
 use Illuminate\Console\Command;
 
 class XCityIdols extends Command
