+++ app/Jobs/OnejavFetchDailyJob.php
@@ -5,7 +5,7 @@ namespace App\Jobs;
 use App\Models\Onejav;
 use App\Models\XCrawlerLog;
 use App\Services\Crawler\OnejavCrawler;
-use App\Services\OnejavService;
+use App\Services\Jav\OnejavService;
 use Carbon\Carbon;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
