+++ app/Services/Jav/XCityIdolService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services;
+namespace App\Services\Jav;
 
 use App\Jobs\XCityIdolFetchPages;
 use App\Services\Crawler\XCityIdolCrawler;
