+++ tests/Unit/Command/R18ReleasedTest.php
@@ -9,7 +9,7 @@ use App\Services\Client\CrawlerClientResponse;
 use App\Services\Client\Domain\ResponseInterface;
 use App\Services\Client\XCrawlerClient;
 use App\Services\Crawler\R18Crawler;
-use App\Services\R18Service;
+use App\Services\Jav\R18Service;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Support\Facades\Queue;
 use PHPUnit\Framework\MockObject\MockObject;
