+++ tests/Unit/Command/XCityIdolTest.php
@@ -7,8 +7,8 @@ use App\Services\Client\CrawlerClientResponse;
 use App\Services\Client\Domain\ResponseInterface;
 use App\Services\Client\XCrawlerClient;
 use App\Services\Crawler\XCityIdolCrawler;
+use App\Services\Jav\XCityIdolService;
 use App\Services\TemporaryUrlService;
-use App\Services\XCityIdolService;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use PHPUnit\Framework\MockObject\MockObject;
