+++ tests/Unit/Command/XCityVideoTest.php
@@ -8,8 +8,8 @@ use App\Services\Client\Domain\ResponseInterface;
 use App\Services\Client\XCrawlerClient;
 use App\Services\Crawler\XCityIdolCrawler;
 use App\Services\Crawler\XCityVideoCrawler;
+use App\Services\Jav\XCityVideoService;
 use App\Services\TemporaryUrlService;
-use App\Services\XCityVideoService;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use PHPUnit\Framework\MockObject\MockObject;
