+++ tests/Unit/Observers/MovieCreatedTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Unit\Observers;
+
+use Tests\TestCase;
+
+class MovieCreatedTest extends TestCase
+{
+    public function test_movie_created_event()
+    {
+
+    }
+}
